perf(ai-section): dynamically import recharts chart and fix hydration mismatch
- Extract chart into AIAnalyticsChart component with dynamic import (ssr: false)
- Remove recharts (~200kb) from initial bundle, matching existing project pattern
- Replace Math.random() in skeleton with deterministic heights to prevent hydration mismatch
